server1/requirements.txt
flask==3.1.0
google-cloud-vision==3.10.1
google-auth==2.39.0

